+++ NEWS.md
@@ -1,6 +1,7 @@
 # ptrr (development version)
 
 * new templates: site, repo docs pages, xaringan slides
+* switch default font to IBM Plex Sans
 
 # ptrr 0.2.0